rework on buttontemplate and buttonconfig
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@1346 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/chan_sccp.h
@@ -399,7 +399,7 @@ struct sccp_linedevices {
  */
 
 struct sccp_buttonconfig {
-        uint8_t					instance;				/*!< Instance on device */
+        uint32_t				instance;				/*!< Instance on device */
         button_type_t 				type;					/*!< Button type (e.g. line, speeddial, feature, empty) */
         SCCP_LIST_ENTRY(sccp_buttonconfig_t) 	list;					/*!< Button Linked List Entry */
